--- water_filter.py
@@ -1,29 +1,0 @@
-
-import cv2
-import numpy as np
-
-def water_detection_mask(img_path: str, min_area_threshold: int = 500) -> np.ndarray:
-    img = cv2.imread(img_path)
-    hsv = cv2.cvtColor(img, cv2.COLOR_BGR2HSV)
-
-    # Color Range
-    lower_water = np.array([90, 50, 50])
-    upper_water = np.array([140, 255, 255])
-
-    mask_water = cv2.inRange(hsv, lower_water, upper_water)
-
-    # Morphological operations (close small gaps in layer)
-    water_kernel_threshold = 12
-    water_kernel = np.ones((water_kernel_threshold, water_kernel_threshold), np.uint8)
-    closed_water_mask = cv2.morphologyEx(mask_water, cv2.MORPH_CLOSE, water_kernel)
-
-    # Find contours in water segments
-    contours, _ = cv2.findContours(closed_water_mask, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
-
-    # Filter out artifacts (small water areas based on given threshold)
-    filtered_water_mask = np.zeros_like(closed_water_mask)
-    for cnt in contours:
-        if cv2.contourArea(cnt) >= min_area_threshold:
-            cv2.drawContours(filtered_water_mask, [cnt], -1, 255, thickness=cv2.FILLED)
-
-    return (filtered_water_mask > 0).astype(np.uint8)
